automations/deal_hunter/requirements.txt
playwright>=1.42.0
python-telegram-bot>=21.0
anthropic>=0.25.0
python-dotenv>=1.0.0
